Rearrange test files in core text fixtures and move them to correct packages
COSDK-546
