+++ core/src/test/java/com/adyen/checkout/core/analytics/internal/data/remote/api/AnalyticsServiceTest.kt
@@ -9,10 +9,10 @@
 package com.adyen.checkout.core.analytics.internal.data.remote.api
 
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
+import com.adyen.checkout.core.common.LoggingExtension
 import com.adyen.checkout.core.common.internal.api.AdyenApiResponse
 import com.adyen.checkout.core.common.internal.api.HttpClient
 import com.adyen.checkout.core.common.internal.model.EmptyResponse
-import com.adyen.checkout.test.LoggingExtension
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertInstanceOf
 import org.junit.jupiter.api.BeforeEach
