+++ core/src/test/java/com/adyen/checkout/core/components/internal/data/api/DefaultStatusRepositoryTest.kt
@@ -8,10 +8,10 @@
 
 package com.adyen.checkout.core.components.internal.data.api
 
+import com.adyen.checkout.core.common.LoggingExtension
+import com.adyen.checkout.core.common.TestDispatcherExtension
+import com.adyen.checkout.core.common.test
 import com.adyen.checkout.core.components.internal.data.model.StatusResponse
-import com.adyen.checkout.test.LoggingExtension
-import com.adyen.checkout.test.TestDispatcherExtension
-import com.adyen.checkout.test.test
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.cancel
