+++ core/src/testFixtures/java/com/adyen/checkout/core/common/TestDispatcherExtension.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ararat on 6/6/2025.
+ * Created by ararat on 26/8/2025.
  */
 
-package com.adyen.checkout.test
+package com.adyen.checkout.core.common
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
