+++ mbway/src/test/java/com/adyen/checkout/mbway/internal/ui/state/MBWayPaymentComponentStateTest.kt
@@ -13,7 +13,7 @@ import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.components.internal.ui.state.model.ComponentFieldState
 import com.adyen.checkout.core.components.internal.ui.state.model.Validation
-import com.adyen.checkout.test.ui.state.TestFieldTransformerRegistry
+import com.adyen.checkout.core.components.internal.ui.state.transformer.TestFieldTransformerRegistry
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
